+++ src/prompts/constants.ts
@@ -1,0 +1,12 @@
+export const PROMPT_DELIMITER = process.env.PROMPTFOO_PROMPT_SEPARATOR || '---';
+export const VALID_FILE_EXTENSIONS = [
+  '.cjs',
+  '.js',
+  '.json',
+  '.jsonl',
+  '.mjs',
+  '.py',
+  '.txt',
+  '.yml',
+  '.yaml',
+];
