+++ src/prompts/index.ts
@@ -1,0 +1,183 @@
+import * as fs from 'fs';
+import { globSync } from 'glob';
+import * as path from 'path';
+import invariant from 'tiny-invariant';
+import logger from '../logger';
+import type {
+  UnifiedConfig,
+  Prompt,
+  ProviderOptionsMap,
+  TestSuite,
+  ProviderOptions,
+} from '../types';
+import { processJsFile } from './processors/javascript';
+import { processJsonFile } from './processors/json';
+import { processJsonlFile } from './processors/jsonl';
+import { processPythonFile } from './processors/python';
+import { processString } from './processors/string';
+import { processTxtFile } from './processors/text';
+import { processYamlFile } from './processors/yaml';
+import { maybeFilePath, normalizeInput, parsePathOrGlob } from './utils';
+
+export * from '../gradingPrompts';
+
+/**
+ * Reads and maps provider prompts based on the configuration and parsed prompts.
+ * @param config - The configuration object.
+ * @param parsedPrompts - Array of parsed prompts.
+ * @returns A map of provider IDs to their respective prompts.
+ */
+export function readProviderPromptMap(
+  config: Partial<UnifiedConfig>,
+  parsedPrompts: Prompt[],
+): TestSuite['providerPromptMap'] {
+  const ret: Record<string, string[]> = {};
+
+  if (!config.providers) {
+    return ret;
+  }
+
+  const allPrompts = [];
+  for (const prompt of parsedPrompts) {
+    allPrompts.push(prompt.label);
+  }
+
+  if (typeof config.providers === 'string') {
+    return { [config.providers]: allPrompts };
+  }
+
+  if (typeof config.providers === 'function') {
+    return { 'Custom function': allPrompts };
+  }
+
+  for (const provider of config.providers) {
+    if (typeof provider === 'object') {
+      // It's either a ProviderOptionsMap or a ProviderOptions
+      if (provider.id) {
+        const rawProvider = provider as ProviderOptions;
+        invariant(
+          rawProvider.id,
+          'You must specify an `id` on the Provider when you override options.prompts',
+        );
+        ret[rawProvider.id] = rawProvider.prompts || allPrompts;
+        if (rawProvider.label) {
+          ret[rawProvider.label] = rawProvider.prompts || allPrompts;
+        }
+      } else {
+        const rawProvider = provider as ProviderOptionsMap;
+        const originalId = Object.keys(rawProvider)[0];
+        const providerObject = rawProvider[originalId];
+        const id = providerObject.id || originalId;
+        ret[id] = rawProvider[originalId].prompts || allPrompts;
+      }
+    }
+  }
+
+  return ret;
+}
+
+type RawPrompt = Partial<Prompt> & {
+  // The source of the prompt, such as a file path or a string.
+  // Could also be an index in the array or a key in the object.
+  source?: string;
+};
+
+/**
+ * Processes a raw prompt based on its content type and path.
+ * @param prompt - The raw prompt data.
+ * @param basePath - Base path for file resolution.
+ * @param maxRecursionDepth - Maximum recursion depth for globbing.
+ * @returns Promise resolving to an array of processed prompts.
+ */
+export async function processPrompt(
+  prompt: RawPrompt,
+  basePath: string = '',
+  maxRecursionDepth: number = 1,
+): Promise<Prompt[]> {
+  invariant(
+    typeof prompt.raw === 'string',
+    `prompt.raw must be a string, but got ${JSON.stringify(prompt.raw)}`,
+  );
+  if (!maybeFilePath(prompt.raw)) {
+    return processString(prompt);
+  }
+
+  const {
+    extension,
+    functionName,
+    isDirectory,
+    promptPath,
+  }: {
+    extension: string;
+    functionName?: string;
+    isDirectory: boolean;
+    promptPath: string;
+  } = parsePathOrGlob(basePath, prompt.raw);
+
+  if (isDirectory && maxRecursionDepth > 0) {
+    const globbedPath = globSync(promptPath.replace(/\\/g, '/'), {
+      windowsPathsNoEscape: true,
+    });
+    logger.debug(
+      `Expanded prompt ${prompt.raw} to ${promptPath} and then to ${JSON.stringify(globbedPath)}`,
+    );
+    const globbedFilePaths = globbedPath.flatMap((globbedPath) => {
+      const filenames = fs.readdirSync(globbedPath);
+      return filenames.map((filename) => path.join(globbedPath, filename));
+    });
+    const prompts: Prompt[] = [];
+    for (const globbedFilePath of globbedFilePaths) {
+      const processedPrompts = await processPrompt(
+        { raw: globbedFilePath },
+        basePath,
+        maxRecursionDepth - 1,
+      );
+      prompts.push(...processedPrompts);
+    }
+    return prompts;
+  }
+
+  if (extension === '.json') {
+    return processJsonFile(promptPath, prompt);
+  }
+  if (extension === '.jsonl') {
+    return processJsonlFile(promptPath, prompt);
+  }
+  if (['.js', '.cjs', '.mjs'].includes(extension)) {
+    return processJsFile(promptPath, functionName);
+  }
+  if (extension === '.py') {
+    return processPythonFile(promptPath, prompt, functionName);
+  }
+  if (extension === '.txt') {
+    return processTxtFile(promptPath, prompt);
+  }
+  if (['.yml', '.yaml'].includes(extension)) {
+    return processYamlFile(promptPath, prompt);
+  }
+  return [];
+}
+
+/**
+ * Reads and processes prompts from a specified path or glob pattern.
+ * @param promptPathOrGlobs - The path or glob pattern.
+ * @param basePath - Base path for file resolution.
+ * @returns Promise resolving to an array of processed prompts.
+ */
+export async function readPrompts(
+  promptPathOrGlobs: string | (string | Partial<Prompt>)[] | Record<string, string>,
+  basePath: string = '',
+): Promise<Prompt[]> {
+  logger.debug(`Reading prompts from ${JSON.stringify(promptPathOrGlobs)}`);
+
+  const promptPartials: RawPrompt[] = normalizeInput(promptPathOrGlobs);
+  const prompts: Prompt[] = [];
+  for (const prompt of promptPartials) {
+    const promptBatch = await processPrompt(prompt, basePath);
+    if (promptBatch.length === 0) {
+      throw new Error(`There are no prompts in ${JSON.stringify(prompt.raw)}`);
+    }
+    prompts.push(...promptBatch);
+  }
+  return prompts;
+}
