+++ src/prompts/utils.ts
@@ -1,0 +1,142 @@
+import * as fs from 'fs';
+import * as path from 'path';
+import { Prompt } from '../types';
+import { VALID_FILE_EXTENSIONS } from './constants';
+
+/**
+ * Determines if a string is a valid file path.
+ * @param str - The string to check.
+ * @returns True if the string is a valid file path, false otherwise.
+ */
+export function maybeFilePath(str: string): boolean {
+  if (typeof str !== 'string') {
+    throw new Error(`Invalid input: ${JSON.stringify(str)}`);
+  }
+
+  const forbiddenSubstrings = ['\n', 'portkey://', 'langfuse://'];
+  if (forbiddenSubstrings.some((substring) => str.includes(substring))) {
+    return false;
+  }
+
+  return (
+    str.startsWith('file://') ||
+    VALID_FILE_EXTENSIONS.some((ext) => {
+      const tokens = str.split(':'); // str may be file.js:functionName
+      // Checks if the second to last token or the last token ends with the extension
+      return tokens.pop()?.endsWith(ext) || tokens.pop()?.endsWith(ext);
+    }) ||
+    str.charAt(str.length - 3) === '.' ||
+    str.charAt(str.length - 4) === '.' ||
+    str.endsWith('.') ||
+    str.includes('*') ||
+    str.includes('/') ||
+    str.includes('\\')
+  );
+}
+
+type RawPrompt = Partial<Prompt> & {
+  // The source of the prompt, such as a file path or a string.
+  // Could also be an index in the array or a key in the object.
+  source?: string;
+};
+
+/**
+ * Normalizes the input prompt to an array of prompts, rejecting invalid and empty inputs.
+ * @param promptPathOrGlobs - The input prompt.
+ * @returns The normalized prompts.
+ * @throws If the input is invalid or empty.
+ */
+export function normalizeInput(
+  promptPathOrGlobs: string | (string | Partial<Prompt>)[] | Record<string, string>,
+): RawPrompt[] {
+  if (
+    !promptPathOrGlobs ||
+    ((typeof promptPathOrGlobs === 'string' || Array.isArray(promptPathOrGlobs)) &&
+      promptPathOrGlobs.length === 0)
+  ) {
+    throw new Error(`Invalid input prompt: ${JSON.stringify(promptPathOrGlobs)}`);
+  }
+  if (typeof promptPathOrGlobs === 'string') {
+    return [
+      {
+        raw: promptPathOrGlobs,
+      },
+    ];
+  }
+  if (Array.isArray(promptPathOrGlobs)) {
+    return promptPathOrGlobs.map((promptPathOrGlob, index) => {
+      if (typeof promptPathOrGlob === 'string') {
+        return {
+          raw: promptPathOrGlob,
+        };
+      }
+      return {
+        raw: promptPathOrGlob.raw || promptPathOrGlob.id,
+        ...promptPathOrGlob,
+      };
+    });
+  }
+
+  if (typeof promptPathOrGlobs === 'object' && Object.keys(promptPathOrGlobs).length) {
+    /* NOTE: This format is considered legacy and has been deprecated. Example:
+      {
+        'prompts.txt': 'foo1',
+        'prompts.py': 'foo2',
+      }
+      */
+    return Object.entries(promptPathOrGlobs).map(([raw, key]) => ({
+      label: key,
+      raw: raw,
+      source: raw,
+    }));
+  }
+  // numbers, booleans, etc
+  throw new Error(`Invalid input prompt: ${JSON.stringify(promptPathOrGlobs)}`);
+}
+
+/**
+ * Parses a file path or glob pattern to extract function names and file extensions.
+ * Function names can be specified in the filename like this:
+ * prompt.py:myFunction or prompts.js:myFunction.
+ * @param basePath - The base path for file resolution.
+ * @param promptPath - The path or glob pattern.
+ * @returns Parsed details including function name, file extension, and directory status.
+ */
+export function parsePathOrGlob(
+  basePath: string,
+  promptPath: string,
+): {
+  extension: string;
+  functionName?: string;
+  isDirectory: boolean;
+  promptPath: string;
+} {
+  let filePath = path.join(basePath, promptPath);
+  if (filePath.includes('file:')) {
+    filePath = filePath.split('file:')[1];
+  }
+
+  let stats;
+  try {
+    stats = fs.statSync(filePath);
+  } catch (err) {
+    if (process.env.PROMPTFOO_STRICT_FILES) {
+      throw err;
+    }
+  }
+  let filename = path.parse(filePath).base;
+  let functionName: string | undefined;
+
+  if (filename.includes(':')) {
+    const splits = filename.split(':');
+    if (splits[0] && ['.js', '.cjs', '.mjs', '.py'].some((ext) => splits[0].endsWith(ext))) {
+      [filename, functionName] = splits;
+    }
+  }
+  return {
+    extension: path.parse(filename).ext,
+    functionName,
+    isDirectory: stats?.isDirectory() ?? false,
+    promptPath: path.join(basePath, filename),
+  };
+}
